+++ CHANGELOG.md
@@ -7,6 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - VIMEET-2373 - make admin's enumeration not possible (V02 intrusion test)
 
+## [2.82.3] - 2021-03-05
+### Fixed
+- MV-311 - fix max size upload video 300M
+
+## [2.82.2] - 2021-03-02
+### Fixed
+- MV-310 - fix error if desktop notifications are not available (ie on iOS)
+
+## [2.82.1] - 2021-02-18
+### Fixed
+- MV-308 - fix "values doesn't support values of type: START_OBJECT" ES error
+
+## [2.82.0] - 2021-02-16### Updated
+- Update Elasticsearch to 6.8.13
+
+## [2.81.0] - 2021-02-16
+### Added
+- VIMEET-2361 - Add link to prefill user mail on registration
+
+### Removed
+- VIMEET-1811 - Remove can_move_meeting column from db
+
+## [2.80.0] - 2021-02-15
+### Added
+- VIMEET-2189 - Mute speakers
+- VIMEET-2386 - Add company name to api conference call
+
 ## [2.79.1] - 2021-02-05
 ### Fixed
 - Redis query cache namespace
@@ -17,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2319 - Send webinar record to speakers
 - VIMEET-2323 - Link target blank
 - VIMEET-2296 - Add participant locale in participants export
-- VIMEET-2295 - Can import partipant's locale
+- VIMEET-2295 - Can import participant's locale
 
 ### Fixed
 - VIMEET-2322 - Set events to public and transparent to enabled title display on Google calendar (else only busy is shown)